Split Codex adapter provider runtime modules
